test: migrate master-merged integration tests off project_api_body
Merging master brought in three server_integration modules
(max_param_length, payload_size, websocket_param_length) that still
authored their app under test as a project_api_body source string —
the parameter this branch removed. The clean merge left a semantic
break: create_server_project() no longer accepts project_api_body, so
all twelve tests errored on the PR's merge-with-master CI run.

Author the three apps as real modules under integration/apps and point
each test at them via api_module=app_module(...), matching the rest of
the migrated harness.
